Add direct parallel multiplier with all stages pipelined
This version is a direct data path multiplier with pipeline of
every adder stages.
Comparing to the 1 stage pipelined version, this version adopted
a more intuitive and loosening coding style, which is the better
practice if the timing is allowed.
The maximum path delay on xc7z020clg400-1 (14bit multiplier)
is around 2.5ns and hence the maximum speed is under 400MHz.