Add binary parallel multiplier with all stages pipelined
This version is a binary data path multiplier with pipeline of
every adder stages.

Comparing to the direct data path multiplier, the binary data path
implementation is more efficient in resources ultilization simply
because the total data path from input to output is shorter and
less pipeline stages are required, as well as the inputs do not
have to be wastefully pipelined merely out of delaying need.

The maximum path delay on xc7z020clg400-1 (14bit multiplier)
is around 2.5ns and hence the maximum speed is under 400MHz.
Resources ultilization is about 1/2 of that of the direct data
path one concerning FFs.