Add direct parallel multiplier with 1 stage pipeline
This version is a direct data path multiplier with flip-flops inserted
to implement a one stage pipeline.

Besides, this version has a strictly instructed coding style, e.g.,
precise (array) data width declaration. This leaves less space for
synthesizer to optimize the stucture (especially the LUT boolean
functions). It is unnecessary and can lead to negetive effect (worse timing)
because systhesizer will optimize the data not being referred to or not having input.

The maximum path delay on xc7z020clg400-1 is around 7ns and hence
the maximum speed is about 150MHz.